# [Fix #24] Run nameless--after-hack-local-variables just once