@nickdrozd 2018-07-23
As I mentioned in that issue, my understanding of hooks and local variables is hazy, so I might be missing something subtle.  That said, this seems to work locally.