@Malabarba 2018-07-25
Thanks for this!
The only negative side-effect I can see is that if someone has some code that calls this mode non-interactively it will stop working with this change.
Ideally, instead of checking for `(called-interactively-p 'any)`, we would have a way of checking if local variables have already been loaded in this file. Do you know if there's a way of checking for that?