@nickdrozd 2018-07-30
I don't know. There must be something. But whatever the exact solution, I think it will look something like this: adding the hack-locals hook if called from a hook, and changing the display immediately if called directly.